protect relation from destroy

+++ app/Http/Controllers/Backend/ChildCategoryController.php
@@ -6,6 +6,8 @@ use App\DataTables\ChildCategoryDataTable;
 use App\Http\Controllers\Controller;
 use App\Models\Category;
 use App\Models\ChildCategory;
+use App\Models\HomePageSetting;
+use App\Models\Product;
 use App\Models\SubCategory;
 use Illuminate\Http\Request;
 use Str;
@@ -110,6 +112,19 @@ class ChildCategoryController extends Controller
     public function destroy(string $id)
     {
         $childCategory = ChildCategory::findOrFail($id);
+
+        if(Product::where('child_category_id', $childCategory->id)->count() > 0) {
+          return response(['status' => 'error', 'message' => 'This item contains relation. Can\' delete it!']);
+        }
+
+        $homeSettings = HomePageSetting::all();
+        foreach($homeSettings as $item) {
+          $array = json_decode($item->value, true);
+          $collection = collect($array);
+          if($collection->contains('child_category', $childCategory->id)) {
+            return response(['status' => 'error', 'message' => 'This item contains relation. Can\' delete it!']);
+          }
+        }
 
         $childCategory->delete();
